api: refactor validation logic
