+++ api/src/controllers/auth.ts
@@ -8,7 +8,6 @@ import {
   signOptions,
   verifyOptions,
 } from "./jwt_options";
-import { validationResult } from "express-validator";
 import Task from "../models/task";
 
 type AuthRespPayload = {
@@ -29,11 +28,6 @@ export const signup = async (
   res: Response<AuthRespPayload>,
   next: NextFunction
 ) => {
-  const errors = validationResult(req);
-  if (!errors.isEmpty()) {
-    const msg = errors.array({ onlyFirstError: true })[0]?.msg;
-    return next(new HttpError(msg, 400));
-  }
   try {
     const { username, password } = req.body;
     await User.create({ username, password });
@@ -134,12 +128,6 @@ export const updateUser = async (
   res: Response<AuthRespPayload>,
   next: NextFunction
 ) => {
-  const errors = validationResult(req);
-  if (!errors.isEmpty()) {
-    const msg = errors.array({ onlyFirstError: true })[0]?.msg;
-    return next(new HttpError(msg, 400));
-  }
-
   try {
     if (!req.userId) {
       return next(new HttpError("userId not found from Request", 400));
